Fix AddCommand Abstraction to be similar to other commands, to follow SLAP principle better
